Adds client update check functionality
Implements a mechanism for the client to request an update check from the server upon connection.

The client sends its current version to the server, which then determines if an update is available based on test mode status. If an update is available, the server sends a command to the client to download the latest version.